requirements.txt
anyio==3.6.2
click==8.1.3
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
libcst==0.4.9
markdown-it-py==2.1.0
mdurl==0.1.2
mypy-extensions==1.0.0
pydantic==1.10.5
Pygments==2.14.0
python-dateutil==2.8.2
python-multipart==0.0.5
PyYAML==6.0
rich==13.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
strawberry-graphql==0.158.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uvicorn==0.20.0
